Un Taek

+++ app/src/main/java/com/limwoon/musicwriter/user/UserCheck.java
@@ -42,7 +42,7 @@ public class UserCheck {
                 String jwtClaim = token.split("\\.")[1];
                 String decodedJwtClaim = new String(Base64.decode(jwtClaim, Base64.NO_WRAP), "UTF-8");
                 JSONObject decodedJwtClaimJSON = new JSONObject(decodedJwtClaim);
-                int userID = decodedJwtClaimJSON.getInt("userID");
+                long userID = decodedJwtClaimJSON.getLong("userID");
                 String userStrID = decodedJwtClaimJSON.getString("userStrID");
                 String userEmail = decodedJwtClaimJSON.getString("userEmail");
                 String userPicUrl = decodedJwtClaimJSON.getString("userPic_url");
